+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Diskover Web Change Log
 
+## [1.0.2] - 2017-06-20
+### added
+- sort by/order in advanced search
+### changed
+- using elasticsearch scroll api to retrieve file results
+### fixed
+- fatal error when searching for more than 10,000 files
+
 ## [1.0.1] - 2017-06-01
 ### changed
 - wording for quick search drop down items and fileview page